+++ test/Feature/CleanRegex/Match/_stream_first/MatchPatternTest.php
@@ -129,14 +129,10 @@ class MatchPatternTest extends TestCase
 
     public function test_groups_and_offsets()
     {
-        $this->assertSame(2, $this->match()->offsets()->first());
         $this->assertSame('123', $this->match()->group(0)->first());
         $this->assertSame(2, $this->match()->group(0)->offsets()->first());
-        $this->assertSame(2, $this->match()->offsets()->first());
 
-        $this->assertSame(2, $this->match()->offsets()->first());
         $this->assertSame('123', $this->match()->group(0)->stream()->first()->text());
-        $this->assertSame(0, $this->match()->offsets()->keys()->first());
         $this->assertSame(2, $this->match()->group(0)->offsets()->first());
     }